Removed debug statement
This commit should be passing all Calc tests on GS
except for Calc: No Statements which only will pass when
statements are added.
It's also failing Calc: Assorted Errors, but that also
may only work when Track B is implemented as it does not
show on GS what it is testing.